+++ AudioResearch/LotrDataLoading.cs
@@ -1,4 +1,5 @@
-﻿using Audio.Storage;
+﻿using AssetEditor;
+using Audio.Storage;
 using CommonControls.Services;
 
 namespace AudioResearch
