+++ KitbasherEditor/ViewModels/SceneExplorerNodeViews/Rmv2/MaterialGeneralViewModel.cs
@@ -1,164 +1,23 @@
 ﻿using CommonControls.Common;
 using CommonControls.FileTypes.RigidModel;
 using CommonControls.FileTypes.RigidModel.Types;
-using CommonControls.PackFileBrowser;
 using CommonControls.Services;
 using CommunityToolkit.Mvvm.Input;
 using MonoGame.Framework.WpfInterop;
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.ComponentModel;
 using System.Linq;
-using System.Windows;
 using System.Windows.Input;
-using TextureEditor.ViewModels;
-using View3D.Components.Component;
 using View3D.SceneNodes;
 using View3D.Services;
 
 namespace KitbasherEditor.ViewModels.SceneExplorerNodeViews.Rmv2
 {
-    public class MaterialGeneralViewModel : NotifyPropertyChangedImpl
+    public partial class MaterialGeneralViewModel : NotifyPropertyChangedImpl
     {
-        public class TextureViewModel : NotifyPropertyChangedImpl, INotifyDataErrorInfo
-        {
-            PackFileService _packfileService;
-            ApplicationSettingsService _appSettingsService;
-            Rmv2MeshNode _meshNode;
-            public TextureType TexureType { get; private set; }
-
-            bool _useTexture = true;
-            public bool UseTexture { get { return _useTexture; } set { SetAndNotify(ref _useTexture, value); UpdatePreviewTexture(value); } }
-            public string TextureTypeStr { get; set; }
-
-            public string Path
-            {
-                get => _path;
-                set
-                {
-                    if (_path == value)
-                        return;
-
-                    _path = value;
-                    ValidateTexturePath();
-                    NotifyPropertyChanged();
-
-                    UpdateTexturePath(value);
-                }
-            }
-
-            public NotifyAttr<bool> IsVisible { get; set; } = new NotifyAttr<bool>(true);
-
-            private readonly Dictionary<string, List<string>> _errorsByPropertyName = new Dictionary<string, List<string>>();
-            private string _path = "";
-
-            public TextureViewModel(Rmv2MeshNode meshNode, PackFileService packfileService, TextureType texureType, ApplicationSettingsService appSettingsService)
-            {
-                _packfileService = packfileService;
-                _meshNode = meshNode;
-                TexureType = texureType;
-                TextureTypeStr = TexureType.ToString();
-                _path = _meshNode.Material.GetTexture(texureType)?.Path;
-                ValidateTexturePath();
-                _appSettingsService = appSettingsService;
-            }
-            public void Paste()
-            {
-                var path = Clipboard.GetText();
-                if (_packfileService.FindFile(path) == null)
-                {
-                    MessageBox.Show($"Invalid path or path not found {path}", "Attention", MessageBoxButton.OK, MessageBoxImage.Exclamation);
-                    return;
-                }
-
-                UpdateTexturePath(path);
-
-            }
-            public void Preview() => TexturePreviewController.CreateWindow(Path, _packfileService, _meshNode.Geometry);
-
-            public void Browse()
-            {
-                using (var browser = new PackFileBrowserWindow(_packfileService))
-                {
-                    browser.ViewModel.Filter.SetExtentions(new List<string>() { ".dds", ".png", });
-                    if (browser.ShowDialog() == true && browser.SelectedFile != null)
-                    {
-                        try
-                        {
-                            var path = _packfileService.GetFullPath(browser.SelectedFile);
-                            UpdateTexturePath(path);
-                        }
-                        catch
-                        {
-                            UpdatePreviewTexture(false);
-                        }
-                    }
-                }
-            }
-
-            private void ValidateTexturePath()
-            {
-                if (Path == null)
-                    return;
-
-                var isFileFound = true;
-
-                if (string.IsNullOrWhiteSpace(Path) == false)
-                    isFileFound = _packfileService.FindFile(Path) != null;
-
-                if (isFileFound == false && Path.Contains("test_mask.dds"))
-                    isFileFound = true;
-
-                if (isFileFound == false)
-                {
-                    var errorMessage = "Invalid Texture Path!";
-                    _errorsByPropertyName[nameof(Path)] = new List<string>() { errorMessage };
-                }
-                else
-                {
-                    _errorsByPropertyName[nameof(Path)] = null;
-                }
-
-                OnErrorsChanged(nameof(Path));
-            }
-
-            public void Remove()
-            {
-                UseTexture = false;
-                UpdateTexturePath("");
-            }
-
-            void UpdateTexturePath(string newPath)
-            {
-                Path = newPath;
-                _meshNode.UpdateTexture(Path, TexureType);
-            }
-
-            void UpdatePreviewTexture(bool value)
-            {
-                _meshNode.UseTexture(TexureType, value);
-            }
-
-            public IEnumerable GetErrors(string propertyName)
-            {
-                return _errorsByPropertyName.ContainsKey(propertyName) ?
-                    _errorsByPropertyName[propertyName] : null;
-            }
-
-            public bool HasErrors => _errorsByPropertyName.Any();
-            public event EventHandler<DataErrorsChangedEventArgs> ErrorsChanged;
-
-            private void OnErrorsChanged(string propertyName)
-            {
-                ErrorsChanged?.Invoke(this, new DataErrorsChangedEventArgs(propertyName));
-            }
-        }
-
         private readonly KitbasherRootScene _kitbasherRootScene;
         Rmv2MeshNode _meshNode;
-        IComponentManager _componentManager;
         PackFileService _pfs;
         ApplicationSettingsService _applicationSettingsService;
 
@@ -198,9 +57,8 @@ namespace KitbasherEditor.ViewModels.SceneExplorerNodeViews.Rmv2
         public UiVertexFormat VertexType { get { return _meshNode.Geometry.VertexFormat; } set { ChangeVertexType(value); } }
         public IEnumerable<UiVertexFormat> PossibleVertexTypes { get; set; }
 
-        public MaterialGeneralViewModel(KitbasherRootScene kitbasherRootScene, Rmv2MeshNode meshNode, PackFileService pfs, IComponentManager componentManager, ApplicationSettingsService applicationSettings)
+        public MaterialGeneralViewModel(KitbasherRootScene kitbasherRootScene, Rmv2MeshNode meshNode, PackFileService pfs,  ApplicationSettingsService applicationSettings)
         {
-            _componentManager = componentManager;
             _kitbasherRootScene = kitbasherRootScene;
             _meshNode = meshNode;
             _pfs = pfs;
@@ -254,7 +112,6 @@ namespace KitbasherEditor.ViewModels.SceneExplorerNodeViews.Rmv2
 
         void ChangeVertexType(UiVertexFormat newFormat)
         {
-            var scenaManager = _componentManager.GetComponent<SceneManager>();
             var skeletonName = _kitbasherRootScene.Skeleton.SkeletonName;
             _meshNode.Geometry.ChangeVertexType(newFormat, skeletonName);
             NotifyPropertyChanged(nameof(VertexType));
